core/atualizar_google_sheets/requirements.txt
gspread==6.2.0
office365==0.3.15
Office365_REST_Python_Client==2.5.13
Office365_REST_Python_Client==2.5.11
openpyxl==3.1.5
pandas==2.2.3
protobuf==6.30.2
python-dotenv==1.1.0
